requirements.txt
adal==1.0.2
applicationinsights==0.11.4
argcomplete==1.9.4
asn1crypto==0.24.0
astroid==1.6.5
azure-cli-core==2.0.38
azure-cli-nspkg==3.0.2
azure-common==1.1.12
azure-graphrbac==0.40.0
azure-mgmt-authorization==0.50.0
azure-mgmt-nspkg==2.0.0
azure-nspkg==2.0.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.2.2
entrypoints==0.2.3
humanfriendly==4.12.1
idna==2.7
isodate==0.6.0
isort==4.3.4
jmespath==0.9.3
keyring==13.1.0
knack==0.3.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
msrest==0.5.1
msrestazure==0.4.33
oauthlib==2.1.0
paramiko==2.4.1
pyasn1==0.4.3
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pylint==1.9.2
PyNaCl==1.2.1
pyOpenSSL==18.0.0
python-dateutil==2.7.3
PyYAML==3.12
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
tabulate==0.8.2
urllib3==1.23
wrapt==1.10.11
